+++ scripts/upgradeSlave.sh
@@ -62,8 +62,10 @@ if [ "$MY_VERSION" == "1.0.1-stable" ]; then
   rm -rf /home/slave-nafsdm/config.conf
   mv /home/slave-nafsdm/config.conf.temp /home/slave-nafsdm/config.conf
 
-  # 1.2 > forward (init.d file)
-  cp nafsdm/systemconfigs/init.d/nafsdm-slave /etc/init.d/nafsdm-slave
+  # 1.2 > forward (enable on boot file)
+  rm -rf /etc/systemd/system/nafsdm-slave.service
+  cp nafsdm/systemconfigs/nafsdm-slave.service /etc/systemd/system/nafsdm-slave.service
+  /usr/bin/env systemctl enable nafsdm-slave
 
   echo "* Upgrade completed."
   echo "* Script has automatically modified your config to match with the new standards."
@@ -72,8 +74,10 @@ elif [ "$MY_VERSION" == "1.1-stable" ]; then
   rm -rf /home/slave-nafsdm/pythondaemon
   cp nafsdm/slave-nafsdm/pythondaemon /home/slave-nafsdm/pythondaemon -R
 
-  # 1.2 > forward (init.d file)
-  cp nafsdm/systemconfigs/init.d/nafsdm-slave /etc/init.d/nafsdm-slave
+  # 1.2 > forward (enable on boot file)
+  rm -rf /etc/systemd/system/nafsdm-slave.service
+  cp nafsdm/systemconfigs/nafsdm-slave.service /etc/systemd/system/nafsdm-slave.service
+  /usr/bin/env systemctl enable nafsdm-slave
 
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
   rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh
@@ -83,6 +87,8 @@ elif [ "$MY_VERSION" == "1.2-stable" ]; then
   cp nafsdm/slave-nafsdm/pythondaemon /home/slave-nafsdm/pythondaemon -R
 
   # 1.2 > forward (enable on boot file)
+  rm -rf /etc/systemd/system/nafsdm-slave.service
+  cp nafsdm/systemconfigs/nafsdm-slave.service /etc/systemd/system/nafsdm-slave.service
   /usr/bin/env systemctl enable nafsdm-slave
 
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"