+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 1.2.1 (Status: DEVELOPMENT, Branch: development)
 * Development feature: specifiy GitHub branch to use
+* Issue #8 fixed (nafsdm-slave not start on boot)
 * Issue #11 fixed (nafsdmctl editor open causing slaves to crash).
 * Issue #12 fixed (CTRL+C in nafsdmctl edit).